Refactor: move route types and function to its own file
