+++ internal/js/modules/k6/browser/common/http.go
@@ -708,58 +708,3 @@ func (r *Response) Text() (string, error) {
 func (r *Response) URL() string {
 	return r.url
 }
-
-// Route allows to handle a request.
-type Route struct {
-	logger         *log.Logger
-	networkManager *NetworkManager
-
-	request *Request
-	handled bool
-}
-
-// NewRoute creates a new Route that allows to modify a request's behavior.
-func NewRoute(logger *log.Logger, networkManager *NetworkManager, request *Request) *Route {
-	return &Route{
-		logger:         logger,
-		networkManager: networkManager,
-		request:        request,
-		handled:        false,
-	}
-}
-
-func (r *Route) Request() *Request { return r.request }
-
-// Abort aborts the request with the given error code.
-func (r *Route) Abort(errorCode string) {
-	err := r.startHandling()
-	if err != nil {
-		r.logger.Errorf("Route:Abort", "rurl:%s err:%s", r.request.URL(), err)
-		return
-	}
-
-	if errorCode == "" {
-		errorCode = "failed"
-	}
-
-	r.networkManager.AbortRequest(r.request.interceptionID, errorCode)
-}
-
-// Continue continues the request.
-func (r *Route) Continue() {
-	err := r.startHandling()
-	if err != nil {
-		r.logger.Errorf("Route:Continue", "rurl:%s err:%s", r.request.URL(), err)
-		return
-	}
-
-	r.networkManager.ContinueRequest(r.request.interceptionID)
-}
-
-func (r *Route) startHandling() error {
-	if r.handled {
-		return fmt.Errorf("route %s is already handled", r.request.URL())
-	}
-	r.handled = true
-	return nil
-}
